refactor: 경로 이름 수정
